update codeforces template

+++ codeforces/SharkieTemplate.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 
-using i64 = int64_t;
+using i64 = long long;
 
 void solve() {
